Integrated CashDispenser and Printer.
